backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
opencv-python-headless==4.10.0.84
moviepy==1.0.3
rembg==2.0.57
edge-tts==6.1.12
numpy==1.26.4
soundfile==0.12.1
mediapipe==0.10.14
aiofiles==23.2.1
httpx==0.27.2
onnxruntime==1.19.2
scipy==1.14.1
imageio==2.35.1
imageio-ffmpeg==0.5.1
decorator==4.4.2
tqdm==4.66.5
requests==2.32.3